feat: move tests out of new features so that upgrad-downgrade tests run on them too
