+++ go/test/endtoend/reparent/plannedreparent/reparent_test.go
@@ -19,12 +19,14 @@ package plannedreparent
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"vitess.io/vitess/go/mysql"
 	"vitess.io/vitess/go/test/endtoend/cluster"
 	"vitess.io/vitess/go/test/endtoend/reparent/utils"
 	"vitess.io/vitess/go/vt/log"
@@ -438,4 +440,95 @@ func TestReplicationStatus(t *testing.T) {
 	assert.NotEmpty(t, replicationStatus.FilePosition)
 	assert.NotEmpty(t, replicationStatus.Position)
 	assert.NotEmpty(t, replicationStatus.RelayLogPosition)
+}
+
+// TestFullStatus tests that the RPC FullStatus works as intended.
+func TestFullStatus(t *testing.T) {
+	defer cluster.PanicHandler(t)
+	clusterInstance := utils.SetupReparentCluster(t, true)
+	defer utils.TeardownCluster(clusterInstance)
+	tablets := clusterInstance.Keyspaces[0].Shards[0].Vttablets
+	utils.ConfirmReplication(t, tablets[0], []*cluster.Vttablet{tablets[1], tablets[2], tablets[3]})
+
+	// Check that full status gives the correct result for a primary tablet
+	primaryStatus, err := utils.TmcFullStatus(context.Background(), tablets[0])
+	require.NoError(t, err)
+	assert.NotEmpty(t, primaryStatus.ServerUuid)
+	assert.NotEmpty(t, primaryStatus.ServerId)
+	// For a primary tablet there is no replication status
+	assert.Nil(t, primaryStatus.ReplicationStatus)
+	assert.Contains(t, primaryStatus.PrimaryStatus.String(), "vt-0000000101-bin")
+	assert.Equal(t, primaryStatus.GtidPurged, "MySQL56/")
+	assert.False(t, primaryStatus.ReadOnly)
+	assert.True(t, primaryStatus.SemiSyncPrimaryEnabled)
+	assert.True(t, primaryStatus.SemiSyncReplicaEnabled)
+	assert.True(t, primaryStatus.SemiSyncPrimaryStatus)
+	assert.False(t, primaryStatus.SemiSyncReplicaStatus)
+	assert.EqualValues(t, 3, primaryStatus.SemiSyncPrimaryClients)
+	assert.EqualValues(t, 1000000000000000000, primaryStatus.SemiSyncPrimaryTimeout)
+	assert.EqualValues(t, 1, primaryStatus.SemiSyncWaitForReplicaCount)
+	assert.Equal(t, "ROW", primaryStatus.BinlogFormat)
+	assert.Equal(t, "FULL", primaryStatus.BinlogRowImage)
+	assert.Equal(t, "ON", primaryStatus.GtidMode)
+	assert.True(t, primaryStatus.LogReplicaUpdates)
+	assert.True(t, primaryStatus.LogBinEnabled)
+	assert.Contains(t, primaryStatus.Version, "5.7")
+	assert.NotEmpty(t, primaryStatus.VersionComment)
+
+	// Check that full status gives the correct result for a replica tablet
+	replicaStatus, err := utils.TmcFullStatus(context.Background(), tablets[1])
+	require.NoError(t, err)
+	assert.NotEmpty(t, replicaStatus.ServerUuid)
+	assert.NotEmpty(t, replicaStatus.ServerId)
+	assert.Contains(t, replicaStatus.ReplicationStatus.Position, "MySQL56/"+replicaStatus.ReplicationStatus.SourceUuid)
+	assert.EqualValues(t, mysql.ReplicationStateRunning, replicaStatus.ReplicationStatus.IoState)
+	assert.EqualValues(t, mysql.ReplicationStateRunning, replicaStatus.ReplicationStatus.SqlState)
+	assert.Equal(t, fileNameFromPosition(replicaStatus.ReplicationStatus.FilePosition), fileNameFromPosition(primaryStatus.PrimaryStatus.FilePosition))
+	assert.LessOrEqual(t, rowNumberFromPosition(replicaStatus.ReplicationStatus.FilePosition), rowNumberFromPosition(primaryStatus.PrimaryStatus.FilePosition))
+	assert.Equal(t, replicaStatus.ReplicationStatus.RelayLogSourceBinlogEquivalentPosition, primaryStatus.PrimaryStatus.FilePosition)
+	assert.Contains(t, replicaStatus.ReplicationStatus.RelayLogFilePosition, "vt-0000000102-relay")
+	assert.Equal(t, replicaStatus.ReplicationStatus.Position, primaryStatus.PrimaryStatus.Position)
+	assert.Equal(t, replicaStatus.ReplicationStatus.RelayLogPosition, primaryStatus.PrimaryStatus.Position)
+	assert.Empty(t, replicaStatus.ReplicationStatus.LastIoError)
+	assert.Empty(t, replicaStatus.ReplicationStatus.LastSqlError)
+	assert.Equal(t, replicaStatus.ReplicationStatus.SourceUuid, primaryStatus.ServerUuid)
+	assert.EqualValues(t, 0, replicaStatus.ReplicationStatus.ReplicationLagSeconds)
+	assert.False(t, replicaStatus.ReplicationStatus.ReplicationLagUnknown)
+	assert.EqualValues(t, 0, replicaStatus.ReplicationStatus.SqlDelay)
+	assert.False(t, replicaStatus.ReplicationStatus.SslAllowed)
+	assert.False(t, replicaStatus.ReplicationStatus.HasReplicationFilters)
+	assert.False(t, replicaStatus.ReplicationStatus.UsingGtid)
+	assert.True(t, replicaStatus.ReplicationStatus.AutoPosition)
+	assert.Equal(t, replicaStatus.ReplicationStatus.SourceHost, utils.Hostname)
+	assert.EqualValues(t, replicaStatus.ReplicationStatus.SourcePort, tablets[0].MySQLPort)
+	assert.Equal(t, replicaStatus.ReplicationStatus.SourceUser, "vt_repl")
+	assert.Contains(t, replicaStatus.PrimaryStatus.String(), "vt-0000000102-bin")
+	assert.Equal(t, replicaStatus.GtidPurged, "MySQL56/")
+	assert.True(t, replicaStatus.ReadOnly)
+	assert.False(t, replicaStatus.SemiSyncPrimaryEnabled)
+	assert.True(t, replicaStatus.SemiSyncReplicaEnabled)
+	assert.False(t, replicaStatus.SemiSyncPrimaryStatus)
+	assert.True(t, replicaStatus.SemiSyncReplicaStatus)
+	assert.EqualValues(t, 0, replicaStatus.SemiSyncPrimaryClients)
+	assert.EqualValues(t, 1000000000000000000, replicaStatus.SemiSyncPrimaryTimeout)
+	assert.EqualValues(t, 1, replicaStatus.SemiSyncWaitForReplicaCount)
+	assert.Equal(t, "ROW", replicaStatus.BinlogFormat)
+	assert.Equal(t, "FULL", replicaStatus.BinlogRowImage)
+	assert.Equal(t, "ON", replicaStatus.GtidMode)
+	assert.True(t, replicaStatus.LogReplicaUpdates)
+	assert.True(t, replicaStatus.LogBinEnabled)
+	assert.Contains(t, replicaStatus.Version, "5.7")
+	assert.NotEmpty(t, replicaStatus.VersionComment)
+}
+
+// fileNameFromPosition gets the file name from the position
+func fileNameFromPosition(pos string) string {
+	return pos[0 : len(pos)-4]
+}
+
+// rowNumberFromPosition gets the row number from the position
+func rowNumberFromPosition(pos string) int {
+	rowNumStr := pos[len(pos)-4:]
+	rowNum, _ := strconv.Atoi(rowNumStr)
+	return rowNum
 }
